Adicionado anotações hibernate e configurado hibernate.cfg
