+++ pizzaria/src/br/com/sliceitup/vo/ClienteVO.java
@@ -1,10 +1,28 @@
 package br.com.sliceitup.vo;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "cliente")
 public class ClienteVO {
 
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "id")
 	private Integer id;
+
+	@Column(name = "nome")
 	private String nome;
+
+	@Column(name = "telefone")
 	private String telefone;
+
+	@Column(name = "endereco")
 	private String endereco;
 
 	public Integer getId() {
