+++ pizzaria/src/br/com/sliceitup/vo/PedidoVO.java
@@ -1,18 +1,43 @@
 package br.com.sliceitup.vo;
 
+import java.sql.Date;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+
+@Entity
+@Table(name="pedido")
 public class PedidoVO {
 
-	private Ineger id;
-	private ClienteVO cliente;
+	@Id
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@Column(name="id")
+	private Integer id;
+	
+	@Column(name="saida")
 	private Date saida;
+
+	@Column(name="data")
 	private Date data;
+	
+	@Column(name="status")
 	private String status;
 
-	public Ineger getId() {
+	@ManyToOne
+	@JoinColumn(name="idcliente")
+	private ClienteVO cliente;
+	
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Ineger id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
