+++ pizzaria/src/br/com/sliceitup/vo/SaborVO.java
@@ -1,8 +1,22 @@
 package br.com.sliceitup.vo;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Entity
+@Table(name = "sabor")
 public class SaborVO {
 
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "id")
 	private Integer id;
+
+	@Column(name = "descricao")
 	private String descricao;
 
 	public Integer getId() {
