+++ pizzaria/src/hibernate.cfg.xml
@@ -1,0 +1,31 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<!DOCTYPE hibernate-configuration PUBLIC
+          "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
+          "http://hibernate.sourceforge.net/hibernate-configuration-3.0.dtd">
+
+<hibernate-configuration>
+
+	<session-factory>
+		<property name="connection.url">
+			jdbc:postgresql://localhost:5432/pizzaria
+		</property>
+		<property name="connection.username">allan</property>
+		<property name="connection.password">postgres</property>
+		<property name="connection.driver_class">
+			org.postgresql.Driver
+		</property>
+		<property name="dialect">
+			org.hibernate.dialect.PostgresPlusDialect
+		</property>
+		<property name="hibernate.show_sql">true</property>
+		<property name="hibernate.format_sql">true</property>
+
+		<!-- Classes Mapeadas -->
+		<mapping class="br.com.sliceitup.vo.ClienteVO"/>
+		<mapping class="br.com.sliceitup.vo.PedidoVO"/>
+		<mapping class="br.com.sliceitup.vo.ItemPedidoVO"/>
+		<mapping class="br.com.sliceitup.vo.SaborItemPedidoVO"/>
+		<mapping class="br.com.sliceitup.vo.SaborVO"/>
+	</session-factory>
+
+</hibernate-configuration>
